agent/requirements.txt
a2a-sdk>=0.3.0
google-genai>=0.3.0
fastapi>=0.115.2
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
openai>=1.0.0
litellm>=1.0.0
